refactor(debounce.ts): moved the debounce hook to its own file
Moved the debounce hook into its own file to allow it to be reused

re #1840
